+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added batch command option `list [-m]` to list all modules
 - App Arguments for setting a custom logging level (`--trace` for Trace Level and `--debug` for Debug and higher) - #389
 - App Arguments for setting the verbosity level (`--silent` for no verbosity, `--medium` for medium verbosity and `--verbose` for very verbose output) - #376
+- Ability for process components to show their progress in the process widget - #384
 
 ### Removed
 - Removed option to enable dev mode in the GUI application using `-dev`. Use `--dev` instead.